Fix: Change Pull Domain Model to Issue
