+++ PRTracker/Source/PR_Listing/ViewModel/PullTableCellViewModel.swift
@@ -14,7 +14,7 @@ final class PullTableCellViewModel {
     let labelList: Observable<[LabelViewModel]?> = Observable(nil)
     let commentCount: Observable<Int?> = Observable(nil)
     
-    func configureCellData(with data: Pull) {
+    func configureCellData(with data: Issue) {
         // TODO: PRTableCellViewModel 구성하기 (지금은 임시로 title, body, mileStone의 title만 표시하도록 함)
         self.title.value = data.title
         self.content.value = data.body
